+++ src/types/api.ts
@@ -61,8 +61,8 @@ export type ReportQueryType =
   | 'in_outs_securities'
 
 export type AccountSnapshotReportQueryType = 'account_at_start' | 'account_at_end'
-export type IndexedReportQueryType = 'cash_flows' | 'securities_flows'
-export type DetailedReportQueryType = Exclude<ReportQueryType, AccountSnapshotReportQueryType | IndexedReportQueryType>
+export type ArrayReportQueryType = 'cash_flows' | 'securities_flows'
+export type DetailedReportQueryType = Exclude<ReportQueryType, AccountSnapshotReportQueryType | ArrayReportQueryType>
 
 export type ReportTimePeriod = '23:59:59' | '08:40:00'
 
@@ -157,8 +157,8 @@ export type ReportResponseShort<T> = {
   }
 }
 
-export type IndexedReportResponse<T> = {
-  report: Record<string, T>
+export type ArrayReportResponse<T> = {
+  report: T[]
 }
 
 type ReportQueryResultMap = {
@@ -167,8 +167,8 @@ type ReportQueryResultMap = {
   account_at_start: AccountAtStartReport
   account_at_end: AccountAtEndReport
   commissions: ReportResponse<CommissionItem>
-  cash_flows: IndexedReportResponse<CashFlowReportItem>
-  securities_flows: IndexedReportResponse<SecuritiesFlowItem>
+  cash_flows: ArrayReportResponse<CashFlowReportItem>
+  securities_flows: ArrayReportResponse<SecuritiesFlowItem>
   in_outs: ReportResponse<InOutItem>
   in_outs_securities: ReportResponse<InOutSecuritiesItem>
 }